refactor(cars): 자동자에 대한 일급 컬렉션 명칭 변경
- 멤버 변수인 cars와 클래스명이 동일하여 구분하기 위해 변경
- 테스트 코드도 동일 이유로 수정
