+++ src/main/java/racingcar/model/collection/CarGroups.java
@@ -1,14 +1,15 @@
-package racingcar.model;
+package racingcar.model.collection;
 
 import java.util.ArrayList;
 import java.util.List;
+import racingcar.model.Car;
 
-public class Cars {
+public class CarGroups {
 
     private final List<Car> cars;
 
 
-    public Cars(String carNames) {
+    public CarGroups(String carNames) {
         cars = new ArrayList<>();
         for (String carName : carNames.split(",")) {
             cars.add(new Car(carName));
